requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
authlib==1.5.2
itsdangerous==2.2.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.20
PyYAML==6.0.3
reportlab==4.4.10
rich==14.3.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==16.0
